% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrainGene.R
\docType{class}
\name{BrainGene}
\alias{BrainGene}
\alias{.BrainGene}
\title{An S4 class to represent a BrainGene, that is, a TrenaGene with standard gene expression data}
\description{
An S4 class to represent a BrainGene, that is, a TrenaGene with standard gene expression data
}
