% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGene.R
\docType{methods}
\name{getEncodeDHS,TrenaGene-method}
\alias{getEncodeDHS,TrenaGene-method}
\alias{getEncodeDHS}
\title{Get all the dnase hypersensitivity regions in the expansive region covered by the enhancer}
\usage{
\S4method{getEncodeDHS}{TrenaGene}(obj)
}
\arguments{
\item{obj}{An object of class TrenaGene}
}
\description{
Get all the dnase hypersensitivity regions in the expansive region covered by the enhancer
}
