% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGene.R
\docType{methods}
\name{getEnhancers,TrenaGene-method}
\alias{getEnhancers,TrenaGene-method}
\alias{getEnhancers}
\title{Get all the enhancer regions for the gene}
\usage{
\S4method{getEnhancers}{TrenaGene}(obj)
}
\arguments{
\item{obj}{An object of class TrenaGene}
}
\description{
Get all the enhancer regions for the gene
}
