% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGene.R
\docType{methods}
\name{getExpressionMatrixNames,TrenaGene-method}
\alias{getExpressionMatrixNames,TrenaGene-method}
\alias{getExpressionMatrixName}
\alias{loadExpressionData,TrenaGene-method}
\title{Get the names of the expression matrices - their names with directory and .RData suffix stripped out}
\usage{
\S4method{getExpressionMatrixNames}{TrenaGene}(obj)

\S4method{loadExpressionData}{TrenaGene}(obj, datasetName)
}
\arguments{
\item{obj}{An object of class TrenaGene}

\item{datasetName}{A numeric matrix}

\item{obj}{An object of class TrenaGene}
}
\description{
Get the names of the expression matrices - their names with directory and .RData suffix stripped out

Get the a specifc expression matrix
}
