% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGene.R
\docType{methods}
\name{getGeneData,TrenaGene-method}
\alias{getGeneData,TrenaGene-method}
\alias{getGeneData}
\title{Get the (typically project-specific) GeneData object}
\usage{
\S4method{getGeneData}{TrenaGene}(obj)
}
\arguments{
\item{obj}{An object of class TrenaGene}
}
\description{
Get the (typically project-specific) GeneData object
}
