% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGene.R
\docType{methods}
\name{getTranscriptsTable,TrenaGene-method}
\alias{getTranscriptsTable,TrenaGene-method}
\alias{getTranscriptsTable}
\title{Get the transcripts for the gene}
\usage{
\S4method{getTranscriptsTable}{TrenaGene}(obj)
}
\arguments{
\item{obj}{An object of class TrenaGene}
}
\description{
Get the transcripts for the gene
}
