% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaGene.R
\docType{class}
\name{TrenaGene-class}
\alias{TrenaGene-class}
\alias{.TrenaGene}
\alias{TrenaGene}
\title{TrenaGene}
\usage{
TrenaGene(trenaGeneData)
}
\arguments{
\item{geneSymbol}{A character string in standard HUGO nomenclature}

\item{genomeName}{A string indicating the genome used by the Trena object.
Currently, only human and mouse ("hg38","mm10") are supported}

\item{expressionDataDirectory}{A string pointing to a collection of RData expression matrices}

\item{quiet}{A logical indicating whether or not the Trena object should print output}
}
\value{
An object of the TrenaGene class
}
\description{
TrenaGene and its (projected) subclasses provide convenient containers in which to collect
 trena-related aggregation of a gene's (a hierarchy of classes) including expression data,
transcript and variant info, genomic and epigenomic context, trena models and/or the means to create them
}
\examples{
# Create a Trena object using the human hg38 genome
cola1a <- TrenaGene("COL1A1", "hg38")

}
